Fix off-by-one error calculating the freelist offset.
The Traffic Server struct VolHeaderFooter ends with a uint16_t[1]
element to stand in for the dynamically-sized freelist. This is why it
subtracts 1 from the segment count when it later calculates the size of
the freelist on disk.

Since we don't do the dumb struct hack, we also don't need the dumb
arithmetic hack.

Signed-off-by: James Peach <jpeach@apache.org>